+++ mycodo/widgets/widget_graph_synchronous.py
@@ -269,7 +269,7 @@ WIDGET_INFORMATION = {
             'type': 'float',
             'default_value': 90.0,
             'constraints_pass': constraints_pass_positive_value,
-            'name': 'Refresh (Seconds)',
+            'name': '{} ({})'.format(lazy_gettext("Refresh"), lazy_gettext("Seconds")),
             'phrase': 'The period of time between refreshing the widget'
         },
         {